EmPower1Chain’s DPC README.md
A Decentralized P2P Community created to support the EmPower1Chain Ecosystem and EmPower it’s Web3 adoption across the known and unknown metaverse. The heart and soul of the EmPower1Chain Organization; carrying out all tasks required to support the blockchain in an ever-changing, web3 ecosystem. The future Leadership, Delegators, Validator, Miners, and Users of the EmPower1Chain’s Blockchain.